[refactor] khushal87 Tile, FeaturedTile
